ai/requirements.txt
fastapi
uvicorn[standard]
pymongo
scikit-learn
pandas
numpy
python-dotenv
httpx